pyproject.toml
[project]
name = "strands-temporal-plugin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.7",
    "strands-agents>=1.6.0",
    "strands-agents-tools>=0.2.5",
    "structlog>=25.4.0",
    "temporalio>=1.16.0",
    "typing-extensions>=4.14.1",
]

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "anyio>=4.10.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.12.11",
]

[tool.ruff]
line-length = 120
extend-include = ["*.ipynb", "**/__init__.py"]
target-version = "py313"
exclude = [
    ".venv",
    "venv",
    "**/__pycache__",
    "**/node_modules",
    "**/.mypy_cache",
    "**/.pytest_cache",
    "**/.ruff_cache",
    "**/.coverage",
    "**/dist",
    "**/build",
    "**/env",
    "**/.venv",
    "**/.ipynb_checkpoints",
]
force-exclude = true

[tool.ruff.lint]
exclude = ["__init__.py", "**/__init__.py"]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "C901", # too complex
    "E501", # line too long
    "E741", # ambiguous variable name
    "F402", # unused import
    "F823", # missing type stubs
    "D100", # missing docstring in public module
    "D106", # missing docstring in public function
]

# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

[tool.ruff.lint.isort]
lines-after-imports = 2
no-sections = true
combine-as-imports = true
split-on-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"**/*.ipynb" = ["F704"]
"tests/**/*.py" = ["D100", "D103"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.mypy]
strict = true
exclude = [
    "venv",
    ".venv",
    ".*/__pycache__",
    ".*/node_modules",
    ".*/.mypy_cache",
    ".*/.pytest_cache",
    ".*/.ruff_cache",
    ".*/.coverage",
    ".*/dist",
]

[tool.pyright]
include = ["src"]
exclude = [
    "venv",
    ".venv",
    "**/__pycache__",
    "**/node_modules",
    "**/.mypy_cache",
    "**/.pytest_cache",
    "**/.ruff_cache",
    "**/.coverage",
    "**/dist",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=term-missing --cov-report=xml --cov-fail-under=85 -n auto"
filterwarnings = [
    "ignore:unclosed database in <sqlite3.Connection object at:ResourceWarning",
]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
skip_covered = true
show_missing = true
precision = 2
exclude_lines = [
    'def __repr__',
    'pragma= no cover',
    'raise NotImplementedError',
    'if TYPE_CHECKING=',
    'if typing.TYPE_CHECKING=',
    '@overload',
    '@typing.overload',
    '\(Protocol\)=$',
    'typing.assert_never',
    'assert_never',
    'if __name__ == .__main__.=',
]

[tool.coverage.paths]
include = ["src/**/*.py"]

[tool.bandit]
exclude_dirs = ["venv", "tests"]
